+++ src/tasks/LoopCharacteristics.js
@@ -3,12 +3,13 @@ import {cloneContent, cloneMessage, unshiftParent} from '../messageHelper';
 
 export default function LoopCharacteristics(activity, loopCharacteristics) {
   const {id, broker, environment} = activity;
+  const {batchSize = 50} = environment.settings;
   const {type = 'LoopCharacteristics', behaviour = {}} = loopCharacteristics;
   const {isSequential = false, collection: collectionExpression, elementVariable = 'item'} = behaviour;
 
   let completionCondition, startCondition, loopCardinality;
   if ('loopCardinality' in behaviour) loopCardinality = behaviour.loopCardinality;
-  if ('loopMaximum' in behaviour) loopCardinality = behaviour.loopMaximum;
+  else if ('loopMaximum' in behaviour) loopCardinality = behaviour.loopMaximum;
 
   if (behaviour.loopCondition) {
     if (behaviour.testBefore) startCondition = behaviour.loopCondition;
@@ -19,7 +20,6 @@ export default function LoopCharacteristics(activity, loopCharacteristics) {
   }
 
   const loopType = getLoopType();
-
   if (!loopType) return;
 
   const {debug} = environment.Logger(type.toLowerCase());
@@ -54,9 +54,13 @@ export default function LoopCharacteristics(activity, loopCharacteristics) {
     if (!executeMessage) throw new Error('LoopCharacteristics execution requires message');
     const {routingKey: executeRoutingKey, redelivered: isRedelivered} = executeMessage.fields || {};
     const {executionId: parentExecutionId} = executeMessage.content;
-    getCharacteristics();
+    if (!getCharacteristics()) return;
 
-    return isSequential ? executeSequential() : executeParallel();
+    try {
+      return isSequential ? executeSequential() : executeParallel();
+    } catch (err) {
+      return activity.emitFatal(new ActivityError(err.message, executeMessage, err), executeMessage.content);
+    }
 
     function executeSequential() {
       let startIndex = 0;
@@ -94,10 +98,6 @@ export default function LoopCharacteristics(activity, loopCharacteristics) {
 
       function onCompleteMessage(_, message) {
         const {content} = message;
-
-        if (content.isRootScope) return;
-        if (!content.isMultiInstance) return;
-
         const loopOutput = getCharacteristics().output;
         if (content.output !== undefined) loopOutput[content.index] = content.output;
 
@@ -125,41 +125,92 @@ export default function LoopCharacteristics(activity, loopCharacteristics) {
     }
 
     function executeParallel() {
+      const {cardinality, getContent: getStartContent} = getCharacteristics();
+
+      if (cardinality === 0) return complete();
+      if (!cardinality) return activity.emitFatal(new ActivityError(`<${id}> cardinality or collection is required in parallel loops`, executeMessage), getStartContent());
+
+      let index = 0, running = 0;
+      if (isRedelivered) {
+        if (!isNaN(executeMessage.content.index)) index = executeMessage.content.index;
+        if (!isNaN(executeMessage.content.running)) running = executeMessage.content.running;
+      }
       subscribe(onCompleteMessage);
+
       if (isRedelivered) return;
 
-      let index = 0, startContent;
-      while ((startContent = next(index))) {
-        debug(`<${parentExecutionId} (${id})> start parallel iteration index ${index}`);
-        broker.publish('execution', 'execute.start', {...startContent, keep: true});
-        index++;
+      return startBatch();
+
+      function startBatch() {
+        const {output: loopOutput, getContent} = getCharacteristics();
+        const batch = [];
+
+        let startContent = next(index);
+        do {
+          debug(`<${parentExecutionId} (${id})> start parallel iteration index ${index}`);
+          batch.push(startContent);
+          running++;
+          index++;
+
+          if (index >= cardinality || running >= batchSize) {
+            break;
+          }
+        } while ((startContent = next(index)));
+
+        broker.publish('execution', 'execute.iteration.batch', {
+          ...getContent(),
+          index,
+          running,
+          output: loopOutput,
+          preventComplete: true,
+        });
+
+        for (const content of batch) {
+          broker.publish('execution', 'execute.start', content);
+        }
       }
 
       function onCompleteMessage(_, message) {
         const {content} = message;
-        if (content.isRootScope) return broker.cancel(executeConsumerTag);
-        if (!content.isMultiInstance) return;
-
-        const loopOutput = getCharacteristics().output;
+        const {output: loopOutput} = getCharacteristics();
         if (content.output !== undefined) loopOutput[content.index] = content.output;
+
+        running--;
 
         broker.publish('execution', 'execute.iteration.completed', {
           ...content,
           ...getCharacteristics().getContent(),
-          index: content.index,
+          index,
+          running,
           output: loopOutput,
           state: 'iteration.completed',
+          preventComplete: true,
         });
 
-        if (isConditionMet(completionCondition, message)) {
-          stop();
-
-          return broker.publish('execution', 'execute.completed', {
-            ...content,
-            ...getCharacteristics().getContent(),
-            output: getCharacteristics().output,
-          });
+        if (running <= 0 && !next(index)) {
+          return complete(content);
         }
+
+        if (isConditionMet(completionCondition, message)) {
+          return complete(content);
+        }
+
+        if (running <= 0) {
+          running = 0;
+          startBatch();
+        }
+      }
+
+      function complete(content) {
+        stop();
+
+        const {getContent, output} = getCharacteristics();
+
+        return broker.publish('execution', 'execute.completed', {
+          ...content,
+          ...getContent(),
+          output,
+        });
       }
     }
 
@@ -193,8 +244,8 @@ export default function LoopCharacteristics(activity, loopCharacteristics) {
     function getCharacteristics() {
       if (loopSettings) return loopSettings;
 
-      const cardinality = getCardinality();
       const collection = getCollection();
+      const cardinality = getCardinality(collection);
 
       const messageContent = {
         ...cloneContent(executeMessage.content),
@@ -202,6 +253,10 @@ export default function LoopCharacteristics(activity, loopCharacteristics) {
         isSequential,
         output: undefined,
       };
+
+      if (cardinality !== undefined && isNaN(cardinality) || cardinality < 0) {
+        return activity.emitFatal(new ActivityError(`<${id}> invalid loop cardinality >${cardinality}<`, executeMessage), messageContent);
+      }
 
       const output = executeMessage.content.output || [];
 
@@ -221,17 +276,13 @@ export default function LoopCharacteristics(activity, loopCharacteristics) {
       return loopSettings;
     }
 
-    function getCardinality() {
-      if (!loopCardinality) return;
-      let value = loopCardinality;
-      if (!value) return;
-
-      value = environment.resolveExpression(value, executeMessage);
-
-      const nValue = Number(value);
-      if (isNaN(nValue)) return activity.emitFatal(new ActivityError(`<${id}> loopCardinality is not a Number >${value}<`, executeMessage));
-
-      return nValue;
+    function getCardinality(collection) {
+      if (!loopCardinality) {
+        return Array.isArray(collection) ? collection.length : undefined;
+      }
+      const value = environment.resolveExpression(loopCardinality, executeMessage);
+      if (value === undefined) return;
+      return Number(value);
     }
 
     function getCollection() {
@@ -240,9 +291,18 @@ export default function LoopCharacteristics(activity, loopCharacteristics) {
       return environment.resolveExpression(collectionExpression, executeMessage);
     }
 
-    function subscribe(onCompleteMessage) {
-      broker.subscribeOnce('api', `activity.*.${parentExecutionId}`, onApiMessage, {consumerTag: apiConsumerTag}, {priority: 400});
-      broker.subscribeTmp('execution', 'execute.completed', onCompleteMessage, {noAck: true, consumerTag: executeConsumerTag, priority: 300});
+    function subscribe(onIterationCompleteMessage) {
+      broker.subscribeTmp('api', `activity.*.${parentExecutionId}`, onApiMessage, {noAck: true, consumerTag: apiConsumerTag}, {priority: 400});
+      broker.subscribeTmp('execution', 'execute.*', onComplete, {noAck: true, consumerTag: executeConsumerTag, priority: 300});
+
+      function onComplete(routingKey, message, ...args) {
+        if (!message.content.isMultiInstance) return;
+        switch (routingKey) {
+          case 'execute.cancel':
+          case 'execute.completed':
+            return onIterationCompleteMessage(routingKey, message, ...args);
+        }
+      }
     }
   }
 
@@ -260,10 +320,9 @@ export default function LoopCharacteristics(activity, loopCharacteristics) {
     broker.cancel(apiConsumerTag);
   }
 
-
   function isConditionMet(condition, message, loopOutput) {
     if (!condition) return false;
-    const testContext = {...cloneMessage(message), loopOutput};
+    const testContext = cloneMessage(message, {loopOutput});
     return environment.resolveExpression(condition, testContext);
   }
 }